MP-195 :truck: Rename: ScanGlobalService를 global 하위로 이동
